Store user's choices in local storage and apply them on load
Previously, each time the documentation was loaded, the choices that
the user had made were lost and the primary item in each switch would
be selected.

For each unique combination of choices, this commit store the user's
selection in local storage and reinstates it when the page is loaded.
The identifier for each combination of choices is determined by
taking the text for each choice, lower casing it, and sorting
alphabetically. This means that choices Alpha, Bravo, Charlie and
bravo, ALPHA, Charlie will both result in an identifier of
alpha-bravo-charlie.

Closes gh-25

Co-authored-by: Jay Bryant <jbryant@pivotal.io>