Explicitly set config model to :attributes
Previously, the configprop inline macro relied on the default config
model of :attributes. Occasionaly, this default would not be used and
the macro's value would not be parsed into the attributes map
correctly. This resulted in a failure to identify that a property
should be written as an environment variable.

This commit updates the processors for the macro to explicity configure
their content model to be :attributes rather than relying upon the
default.

Closes gh-51