Work around change in attribute parsing across conversions
There's a bug in AsciidoctorJ that causes an inline macro's attributes
to not be parsed after the first conversion performed by an Asciidoctor
instance. It appears that the content_model for the macro changes after
the first invocation so that it no longer uses :attributes. As a result
the parsing is disabled and the unparsed value is stored under the
text key in the attributes map.

This commit works around the problem by changing all of the inline
macro processors to normalize the attributes into their parsed form
before working with them.

Closes gh-51